Move CoordinatorException to the consensuscommit package
