+++ core/src/main/java/com/scalar/db/transaction/consensuscommit/Coordinator.java
@@ -14,7 +14,6 @@ import com.scalar.db.api.TableMetadata;
 import com.scalar.db.api.TransactionState;
 import com.scalar.db.exception.storage.ExecutionException;
 import com.scalar.db.exception.storage.NoMutationException;
-import com.scalar.db.exception.transaction.CoordinatorException;
 import com.scalar.db.io.DataType;
 import com.scalar.db.io.Key;
 import java.util.Objects;
